+++ src/app/app.component.ts
@@ -3,9 +3,9 @@ import { RouterOutlet } from '@angular/router';
 
 import { ProductsModule } from './products/products.module';
 import { APP_INFO, AppInfo } from './core/services/constants.service';
-import { RANDON_GENERATOR, generatorFactory } from './core/services/generator.factory';
-import { GeneratorService } from './core/services/generator.service';
 import { Subscription, interval, map, timer } from 'rxjs';
+import { AppSettingsService, RANDON_GENERATOR, generatorFactory, GeneratorService } from './core/services';
+import { HttpClientModule } from '@angular/common/http';
 
 @Component({
   selector: 'app-root',
@@ -13,9 +13,11 @@ import { Subscription, interval, map, timer } from 'rxjs';
   imports: [
     ProductsModule,
     RouterOutlet,
+    HttpClientModule,
   ],
   providers: [
     GeneratorService,
+    AppSettingsService,
     {provide: APP_INFO, useValue: AppInfo},
     {provide: RANDON_GENERATOR, useFactory: generatorFactory(6)}
     // {provide: RANDON_GENERATOR, useFactory: generatorFactory(6), deps: [GeneratorService]}
@@ -33,10 +35,16 @@ export class AppComponent implements OnInit, OnDestroy, AfterViewInit {
   constructor(
     @Inject(APP_INFO) public appInfo: {App: string, Ver: string, API_URL: string},
     @Inject(RANDON_GENERATOR) public sequience: string,
+    private appSettings: AppSettingsService,
     @Optional() private generator: GeneratorService) {
   }
 
   ngOnInit(): void {
+
+    this.appSettings.getSettings()
+      .subscribe(settings => {
+        console.log(settings);
+      });
 
     this.timerSubscription$ = interval(1500)
     .pipe(map(_ => this.generator.getNewID()))